+++ app/src/main/java/com/sedilant/yambol/domain/get/GetAllTrainsByTeamIdUseCase.kt
@@ -3,5 +3,5 @@ package com.sedilant.yambol.domain.get
 import com.sedilant.yambol.domain.models.TrainDomainModel
 
 interface GetAllTrainsByTeamIdUseCase {
-    suspend operator fun invoke(teamId: Int): List<TrainDomainModel>
+    suspend operator fun invoke(teamId: Long): List<TrainDomainModel>
 }
