+++ app/src/main/java/com/sedilant/yambol/domain/insert/InsertTeamObjectiveUseCase.kt
@@ -1,5 +1,5 @@
 package com.sedilant.yambol.domain.insert
 
 interface InsertTeamObjectiveUseCase {
-    suspend operator fun invoke(description: String, teamId: Int)
+    suspend operator fun invoke(description: String, teamId: Long)
 }
